tst/testall.g: load GAPDoc and SmallGrp before testing
To help resolve the chicken and egg problem, it's nice to be able to
launch gap with --bare when testing transgrp. But without GAPDoc and
SmallGrp loaded, the test suite fails. So let's load them explicitly.
This won't hurt anything under normal circumstances where both of them
are loaded anyway.